+++ module-article/module-article-model/src/main/java/io/jpress/module/article/model/base/BaseArticleComment.java
@@ -89,12 +89,12 @@ public abstract class BaseArticleComment<M extends BaseArticleComment<M>> extend
 		return getLong("reply_count");
 	}
 
-	public void setOrderNumber(java.lang.Long orderNumber) {
+	public void setOrderNumber(java.lang.Integer orderNumber) {
 		set("order_number", orderNumber);
 	}
 	
-	public java.lang.Long getOrderNumber() {
-		return getLong("order_number");
+	public java.lang.Integer getOrderNumber() {
+		return getInt("order_number");
 	}
 
 	public void setVoteUp(java.lang.Long voteUp) {
